Move AtFwd hal from device to framework manifest.
AtFwd hal is implemented as an android service and
it stars from system partition, that is, it's not
a traditional vendor hal and hence it should be
declared in the framework manifest and not in the
device manifest.
This commit introduces new device specific framework
manifest (framework_manifest.xml) and moves AtFwd hal
from device to framework manifest.

Remove IConfig and IOemHook hals.

Change-Id: Ibe179b243081532fc9b5fc82b7ed43120daf9bf3
CRs-Fixed: 2287149